fix(container): Fix container build and image check logic
This commit refines the container build process by ensuring the `/rootfs` directory is explicitly created within the Containerfile and by correcting the volume declaration.

Additionally, the logic for checking the `TITANOBOA_CTR_IMAGE` variable in `main.sh` has been adjusted to prevent premature exits when running inside the container.

+++ main.sh
@@ -39,11 +39,6 @@
 
     SCRIPT_DIR=$(dirname "$0")
 
-    if [ -z "$TITANOBOA_CTR_IMAGE" ]; then
-        echo "Error: container image in param 1 nor TITANOBOA_CTR_IMAGE environment variable"
-        exit 1
-    fi
-
     mkdir -p "$TITANOBOA_OUTPUT_DIR"
 
     # If we are running inside a container
@@ -63,6 +58,12 @@
         fi
         /app/bin/build_iso.sh
     else
+
+        if [ -z "$TITANOBOA_CTR_IMAGE" ]; then
+            echo "Error: container image in param 1 nor TITANOBOA_CTR_IMAGE environment variable"
+            exit 1
+        fi
+
         sudo podman run --rm -i \
             --cap-add sys_admin --security-opt label=disable \
             -v "$SCRIPT_DIR"/build_iso.sh:/src/build_iso.sh:ro \
